Move text-post queries into src/data/text
The Teia Text queries and their hooks lived in src/data/swr.js, a 1.4k-line
file of unrelated hooks. They landed there with the blog in March, before
the repo settled on per-feature data modules; poll comments, wiki and
curations all since got a src/data/<feature>/ folder of their own.

Text now follows that layout: queries.ts holds the GraphQL documents,
hooks.ts the SWR hooks, and index.ts re-exports both so pages import from
'@data/text'. The hooks keep their SWR keys and options exactly, so caching
and what each feed returns are unchanged.

useMultisigAddresses stays in swr.js since New Post uses it too, and roles.ts
already marks it for consolidation with the role hooks.
